+++ docker.sh
@@ -76,6 +76,7 @@ function buildImages {
     buildImage 1.14 pro php7.3 fpm
     buildImage 1.14 pro php7.4 apache
     buildImage 1.14 pro php7.4 fpm
+
     buildImage 1.14 open php7.1 apache
     buildImage 1.14 open php7.1 fpm
     buildImage 1.14 open php7.2 apache
@@ -84,6 +85,7 @@ function buildImages {
     buildImage 1.14 open php7.3 fpm
     buildImage 1.14 open php7.4 apache
     buildImage 1.14 open php7.4 fpm
+
     buildImage 1.14.1 pro php7.1 apache
     buildImage 1.14.1 pro php7.1 fpm
     buildImage 1.14.1 pro php7.2 apache
@@ -92,6 +94,7 @@ function buildImages {
     buildImage 1.14.1 pro php7.3 fpm
     buildImage 1.14.1 pro php7.4 apache
     buildImage 1.14.1 pro php7.4 fpm
+
     buildImage 1.14.1 open php7.1 apache
     buildImage 1.14.1 open php7.1 fpm
     buildImage 1.14.1 open php7.2 apache
@@ -100,6 +103,7 @@ function buildImages {
     buildImage 1.14.1 open php7.3 fpm
     buildImage 1.14.1 open php7.4 apache
     buildImage 1.14.1 open php7.4 fpm
+
     buildImage 1.14.2 pro php7.1 apache
     buildImage 1.14.2 pro php7.1 fpm
     buildImage 1.14.2 pro php7.2 apache
@@ -108,6 +112,15 @@ function buildImages {
     buildImage 1.14.2 pro php7.3 fpm
     buildImage 1.14.2 pro php7.4 apache
     buildImage 1.14.2 pro php7.4 fpm
+
+    buildImage 1.14.2 open php7.1 apache
+    buildImage 1.14.2 open php7.1 fpm
+    buildImage 1.14.2 open php7.2 apache
+    buildImage 1.14.2 open php7.2 fpm
+    buildImage 1.14.2 open php7.3 apache
+    buildImage 1.14.2 open php7.3 fpm
+    buildImage 1.14.2 open php7.4 apache
+    buildImage 1.14.2 open php7.4 fpm
 }
 
 function pullImages {
@@ -156,6 +169,7 @@ function scanImages {
     scanImage 1.14 pro php7.3 fpm
     scanImage 1.14 pro php7.4 apache
     scanImage 1.14 pro php7.4 fpm
+
     scanImage 1.14 open php7.1 apache
     scanImage 1.14 open php7.1 fpm
     scanImage 1.14 open php7.2 apache
@@ -164,6 +178,7 @@ function scanImages {
     scanImage 1.14 open php7.3 fpm
     scanImage 1.14 open php7.4 apache
     scanImage 1.14 open php7.4 fpm
+
     scanImage 1.14.1 pro php7.1 apache
     scanImage 1.14.1 pro php7.1 fpm
     scanImage 1.14.1 pro php7.2 apache
@@ -172,6 +187,7 @@ function scanImages {
     scanImage 1.14.1 pro php7.3 fpm
     scanImage 1.14.1 pro php7.4 apache
     scanImage 1.14.1 pro php7.4 fpm
+
     scanImage 1.14.1 open php7.1 apache
     scanImage 1.14.1 open php7.1 fpm
     scanImage 1.14.1 open php7.2 apache
@@ -180,6 +196,7 @@ function scanImages {
     scanImage 1.14.1 open php7.3 fpm
     scanImage 1.14.1 open php7.4 apache
     scanImage 1.14.1 open php7.4 fpm
+
     scanImage 1.14.2 pro php7.1 apache
     scanImage 1.14.2 pro php7.1 fpm
     scanImage 1.14.2 pro php7.2 apache
@@ -188,6 +205,15 @@ function scanImages {
     scanImage 1.14.2 pro php7.3 fpm
     scanImage 1.14.2 pro php7.4 apache
     scanImage 1.14.2 pro php7.4 fpm
+
+    scanImage 1.14.2 open php7.1 apache
+    scanImage 1.14.2 open php7.1 fpm
+    scanImage 1.14.2 open php7.2 apache
+    scanImage 1.14.2 open php7.2 fpm
+    scanImage 1.14.2 open php7.3 apache
+    scanImage 1.14.2 open php7.3 fpm
+    scanImage 1.14.2 open php7.4 apache
+    scanImage 1.14.2 open php7.4 fpm
 }
 
 function scanImage {
@@ -394,6 +420,15 @@ function pushImages {
     pushImage 1.14.2 pro php7.3 fpm
     pushImage 1.14.2 pro php7.4 apache
     pushImage 1.14.2 pro php7.4 fpm
+
+    pushImage 1.14.2 open php7.1 apache
+    pushImage 1.14.2 open php7.1 fpm
+    pushImage 1.14.2 open php7.2 apache
+    pushImage 1.14.2 open php7.2 fpm
+    pushImage 1.14.2 open php7.3 apache
+    pushImage 1.14.2 open php7.3 fpm
+    pushImage 1.14.2 open php7.4 apache
+    pushImage 1.14.2 open php7.4 fpm
 }
 
 function pushImage {
@@ -410,6 +445,15 @@ function pushImage {
 }
 
 function printReadme {
+    printSupportedTags 1.14.2 open php7.4 fpm
+    printSupportedTags 1.14.2 open php7.4 apache
+    printSupportedTags 1.14.2 open php7.3 fpm
+    printSupportedTags 1.14.2 open php7.3 apache
+    printSupportedTags 1.14.2 open php7.2 fpm
+    printSupportedTags 1.14.2 open php7.2 apache
+    printSupportedTags 1.14.2 open php7.1 fpm
+    printSupportedTags 1.14.2 open php7.1 apache
+
     printSupportedTags 1.14.2 pro php7.4 fpm
     printSupportedTags 1.14.2 pro php7.4 apache
     printSupportedTags 1.14.2 pro php7.3 fpm